# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DiaryEntryUpsert {
  /**
     * Open-text training, mood, recovery, or life note for the selected day
     * @minLength 1
     * @maxLength 2000
     */
  note: string;
}

export type DiaryEntrySentiment = typeof DiaryEntrySentiment[keyof typeof DiaryEntrySentiment] | null;


export const DiaryEntrySentiment = {
  positive: 'positive',
  neutral: 'neutral',
  negative: 'negative',
} as const;

export type DiaryEntrySentimentStatus = typeof DiaryEntrySentimentStatus[keyof typeof DiaryEntrySentimentStatus];


export const DiaryEntrySentimentStatus = {
  analyzed: 'analyzed',
  unavailable: 'unavailable',
} as const;

export interface DiaryEntry {
  id: string;
  entry_date: string;
  note: string;
  sentiment: DiaryEntrySentiment;
  sentiment_confidence: number | null;
  /** Short non-clinical summary; never a diagnosis */
  sentiment_summary: string | null;
  sentiment_status: DiaryEntrySentimentStatus;
  updated_at: string;
}

export type DiaryTrendPointSentiment = typeof DiaryTrendPointSentiment[keyof typeof DiaryTrendPointSentiment];


export const DiaryTrendPointSentiment = {
  positive: 'positive',
  neutral: 'neutral',
  negative: 'negative',
} as const;

export interface DiaryTrendPoint {
  entry_date: string;
  sentiment: DiaryTrendPointSentiment;
  mean_velocity_ms: number | null;
}

export interface DiaryTrend {
  days: number;
  analyzed_entries: number;
  performance_data_points: number;
  correlation_summary: string;
  points: DiaryTrendPoint[];
}

/**
 * A single three-axis accelerometer reading with a millisecond timestamp
 */
export interface AccelerationSample {
  /** X-axis acceleration in G */
  x: number;
  /** Y-axis acceleration in G */
  y: number;
  /** Z-axis acceleration in G */
  z: number;
  /** Unix timestamp in milliseconds when the sample was captured */
  timestamp: number;
}

/**
 * User-facing mass unit used for load labels and coaching copy
 */
export type SetAnalysisRequestDisplayUnit = typeof SetAnalysisRequestDisplayUnit[keyof typeof SetAnalysisRequestDisplayUnit];


export const SetAnalysisRequestDisplayUnit = {
  imperial: 'imperial',
  metric: 'metric',
} as const;

/**
 * How the phone is mounted during the set. Affects which axis is used for velocity integration and how the AI coaching interprets the data.
 */
export type SetAnalysisRequestPhonePlacement = typeof SetAnalysisRequestPhonePlacement[keyof typeof SetAnalysisRequestPhonePlacement];


export const SetAnalysisRequestPhonePlacement = {
  weight_stack: 'weight_stack',
  barbell: 'barbell',
  pocket: 'pocket',
} as const;

/**
 * Exercise setup metadata and the raw sensor batch captured during a set
 */
export interface SetAnalysisRequest {
  /**
     * Name of the exercise (e.g. "Bench Press")
     * @minLength 1
     */
  exercise_name: string;
  /** Load on the bar in kilograms */
  weight_kg: number;
  /** User-facing mass unit used for load labels and coaching copy */
  display_unit?: SetAnalysisRequestDisplayUnit;
  /** User-entered load in display_unit; weight_kg remains canonical */
  display_load?: number;
  /** Planned rep count for the set */
  target_reps: number;
  /** Total sets planned in the workout */
  total_sets: number;
  /** Ordered raw acceleration samples captured during the set */
  samples: AccelerationSample[];
  /** How the phone is mounted during the set. Affects which axis is used for velocity integration and how the AI coaching interprets the data. */
  phone_placement?: SetAnalysisRequestPhonePlacement;
}

export type SetAnalysisResultStatus = typeof SetAnalysisResultStatus[keyof typeof SetAnalysisResultStatus];


export const SetAnalysisResultStatus = {
  success: 'success',
} as const;

export interface SetAnalysisResult {
  status: SetAnalysisResultStatus;
  exercise_name: string;
  /** Mean bar velocity in m/s integrated from the acceleration batch */
  mean_velocity_ms: number;
  /** Peak bar velocity in m/s */
  peak_velocity_ms: number;
  /** First-rep peak velocity in m/s — primary CNS readiness marker. Null when fewer than 1 rep detected. */
  first_rep_peak_ms: number | null;
  /** Estimated percentage of 1RM based on exercise-specific load-velocity profile */
  estimated_1rm_pct: number;
  /** VBT training zone: Maximal Strength / Strength-Speed / Speed-Strength / Power / Starting */
  velocity_zone: string;
  /** Velocity loss from first to last rep (%). Null when fewer than 2 reps detected. */
  velocity_loss_pct: number | null;
  /** Qualitative fatigue classification based on velocity loss */
  fatigue_level: string | null;
  /** Number of samples processed */
  sample_count: number;
  /** Total set duration in seconds */
  duration_s: number;
  /** AI-generated coaching feedback grounded in VBT standards */
  ai_feedback: string;
  /** Whether historical Sparkden session data was incorporated in the feedback */
  sparkden_history_used: boolean;
  /** Motor readiness score 0–100 based on first-rep peak vs. 21-day load-matched baseline. Null when fewer than 3 historical sets exist at this load. */
  cns_readiness_score: number | null;
  /** Qualitative readiness classification: High / Moderate / Low / Compromised / Insufficient data */
  motor_readiness_level: string | null;
  /** Session-to-session velocity trend: Rising / Stable / Declining / Insufficient data */
  velocity_trend: string | null;
  /** Number of historical load-matched sessions used to compute readiness */
  readiness_data_points: number;
  /** The 21-day load-matched mean first-rep peak used as the readiness baseline. Null when insufficient data. */
  baseline_velocity_ms: number | null;
  /** Number of reps detected from the velocity trace via peak-detection */
  actual_reps: number;
  /** Per-rep peak velocity in m/s, ordered first rep to last rep */
  rep_peaks_ms: number[];
  /** Actionable comparison against the selected exercise's load-matched historical profile */
  historical_comparison: string;
  /** Current mean velocity deviation from the load-matched historical baseline (%) */
  historical_comparison_delta_pct: number | null;
  /** Load-matched historical mean-velocity baseline in m/s */
  historical_baseline_velocity_ms: number | null;
  /** Number of historical load-matched sessions used for the comparison */
  historical_comparison_data_points: number;
  /** Whether an optional diary sentiment context was available for this set date */
  diary_context_available: boolean;
  /** Non-clinical sentiment context only; does not contain the raw diary note */
  diary_context: string | null;
}

/**
 * Immutable demo history row. Loads are always canonical kilograms.
 */
export interface DemoHistoricalSet {
  id: string;
  setNumber: number;
  exercise: string;
  daysBeforeDemo: number;
  /** Canonical load in kilograms */
  loadKg: number;
  targetReps: number;
  actualReps: number;
  meanRepTimeSec: number;
  velocityLossPct: number;
  displacementM: number | null;
  source: string;
  provenance: string;
}

export type GetDiaryEntriesParams = {
start_date: string;
end_date: string;
};

export type GetDiaryTrendParams = {
/**
 * @minimum 7
 * @maximum 90
 */
days?: number;
};