# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: coaching
    description: Set analysis operations
  - name: diary
    description: Daily training notes and non-clinical sentiment context
paths:
  /:
    get:
      operationId: apiHome
      tags: [health]
      summary: API welcome message
      description: Confirms that the Velocity Unlimited API is live
      responses:
        "200":
          description: API is live
          content:
            text/plain:
              schema:
                type: string
  /analyze-set:
    post:
      operationId: analyzeSet
      tags: [coaching]
      summary: Analyze a set
      description: Accepts a raw acceleration batch and returns velocity-integrated coaching feedback enriched with VBT profiles, CNS readiness score, and AI coaching text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAnalysisRequest"
      responses:
        "200":
          description: Coaching feedback with velocity metrics and CNS readiness data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetAnalysisResult"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /history:
    get:
      operationId: getDemoHistory
      tags: [coaching]
      summary: Get canonical demo history
      description: Returns the immutable 24-set demo history in canonical kilograms
      responses:
        "200":
          description: Canonical demo history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DemoHistoricalSet"
  /diary:
    get:
      operationId: getDiaryEntries
      tags: [diary]
      summary: Get diary entries in a date range
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Diary entries ordered by date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiaryEntry"
  /diary/{date}:
    get:
      operationId: getDiaryEntry
      tags: [diary]
      summary: Get a diary entry for a day
      parameters:
        - $ref: "#/components/parameters/DiaryDate"
      responses:
        "200":
          description: Entry for the requested date, or null when none exists
          content:
            application/json:
              schema:
                nullable: true
                allOf:
                  - $ref: "#/components/schemas/DiaryEntry"
    put:
      operationId: saveDiaryEntry
      tags: [diary]
      summary: Create or update a diary entry
      parameters:
        - $ref: "#/components/parameters/DiaryDate"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiaryEntryUpsert"
      responses:
        "200":
          description: Saved entry with sentiment availability details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryEntry"
  /diary-trend:
    get:
      operationId: getDiaryTrend
      tags: [diary]
      summary: Get diary and performance trend context
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 28
            minimum: 7
            maximum: 90
      responses:
        "200":
          description: Non-clinical diary and performance overlap summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryTrend"
components:
  parameters:
    DiaryDate:
      name: date
      in: path
      required: true
      schema:
        type: string
        format: date
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DiaryEntryUpsert:
      type: object
      properties:
        note:
          type: string
          minLength: 1
          maxLength: 2000
          description: Open-text training, mood, recovery, or life note for the selected day
      required:
        - note
    DiaryEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entry_date:
          type: string
          format: date
        note:
          type: string
        sentiment:
          type: string
          nullable: true
          enum: [positive, neutral, negative, null]
        sentiment_confidence:
          type: number
          nullable: true
        sentiment_summary:
          type: string
          nullable: true
          description: Short non-clinical summary; never a diagnosis
        sentiment_status:
          type: string
          enum: [analyzed, unavailable]
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - entry_date
        - note
        - sentiment
        - sentiment_confidence
        - sentiment_summary
        - sentiment_status
        - updated_at
    DiaryTrendPoint:
      type: object
      properties:
        entry_date:
          type: string
          format: date
        sentiment:
          type: string
          enum: [positive, neutral, negative]
        mean_velocity_ms:
          type: number
          nullable: true
      required:
        - entry_date
        - sentiment
        - mean_velocity_ms
    DiaryTrend:
      type: object
      properties:
        days:
          type: integer
        analyzed_entries:
          type: integer
        performance_data_points:
          type: integer
        correlation_summary:
          type: string
        points:
          type: array
          items:
            $ref: "#/components/schemas/DiaryTrendPoint"
      required:
        - days
        - analyzed_entries
        - performance_data_points
        - correlation_summary
        - points
    AccelerationSample:
      type: object
      description: A single three-axis accelerometer reading with a millisecond timestamp
      properties:
        x:
          type: number
          description: X-axis acceleration in G
        y:
          type: number
          description: Y-axis acceleration in G
        z:
          type: number
          description: Z-axis acceleration in G
        timestamp:
          type: number
          description: Unix timestamp in milliseconds when the sample was captured
      required:
        - x
        - y
        - z
        - timestamp
    SetAnalysisRequest:
      type: object
      description: Exercise setup metadata and the raw sensor batch captured during a set
      properties:
        exercise_name:
          type: string
          minLength: 1
          description: Name of the exercise (e.g. "Bench Press")
        weight_kg:
          type: number
          description: Load on the bar in kilograms
        display_unit:
          type: string
          enum: [imperial, metric]
          default: metric
          description: User-facing mass unit used for load labels and coaching copy
        display_load:
          type: number
          description: User-entered load in display_unit; weight_kg remains canonical
        target_reps:
          type: integer
          description: Planned rep count for the set
        total_sets:
          type: integer
          description: Total sets planned in the workout
        samples:
          type: array
          description: Ordered raw acceleration samples captured during the set
          items:
            $ref: "#/components/schemas/AccelerationSample"
        phone_placement:
          type: string
          enum: [weight_stack, barbell, pocket]
          default: weight_stack
          description: "How the phone is mounted during the set. Affects which axis is used for velocity integration and how the AI coaching interprets the data."
      required:
        - exercise_name
        - weight_kg
        - target_reps
        - total_sets
        - samples
    SetAnalysisResult:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        exercise_name:
          type: string
        mean_velocity_ms:
          type: number
          description: Mean bar velocity in m/s integrated from the acceleration batch
        peak_velocity_ms:
          type: number
          description: Peak bar velocity in m/s
        first_rep_peak_ms:
          type: number
          description: "First-rep peak velocity in m/s — primary CNS readiness marker. Null when fewer than 1 rep detected."
          nullable: true
        estimated_1rm_pct:
          type: integer
          description: Estimated percentage of 1RM based on exercise-specific load-velocity profile
        velocity_zone:
          type: string
          description: "VBT training zone: Maximal Strength / Strength-Speed / Speed-Strength / Power / Starting"
        velocity_loss_pct:
          type: number
          description: "Velocity loss from first to last rep (%). Null when fewer than 2 reps detected."
          nullable: true
        fatigue_level:
          type: string
          description: "Qualitative fatigue classification based on velocity loss"
          nullable: true
        sample_count:
          type: integer
          description: Number of samples processed
        duration_s:
          type: number
          description: Total set duration in seconds
        ai_feedback:
          type: string
          description: AI-generated coaching feedback grounded in VBT standards
        sparkden_history_used:
          type: boolean
          description: Whether historical Sparkden session data was incorporated in the feedback
        cns_readiness_score:
          type: integer
          description: "Motor readiness score 0–100 based on first-rep peak vs. 21-day load-matched baseline. Null when fewer than 3 historical sets exist at this load."
          nullable: true
        motor_readiness_level:
          type: string
          description: "Qualitative readiness classification: High / Moderate / Low / Compromised / Insufficient data"
          nullable: true
        velocity_trend:
          type: string
          description: "Session-to-session velocity trend: Rising / Stable / Declining / Insufficient data"
          nullable: true
        readiness_data_points:
          type: integer
          description: Number of historical load-matched sessions used to compute readiness
        baseline_velocity_ms:
          type: number
          description: "The 21-day load-matched mean first-rep peak used as the readiness baseline. Null when insufficient data."
          nullable: true
        actual_reps:
          type: integer
          description: Number of reps detected from the velocity trace via peak-detection
        rep_peaks_ms:
          type: array
          description: Per-rep peak velocity in m/s, ordered first rep to last rep
          items:
            type: number
        historical_comparison:
          type: string
          description: Actionable comparison against the selected exercise's load-matched historical profile
        historical_comparison_delta_pct:
          type: number
          nullable: true
          description: Current mean velocity deviation from the load-matched historical baseline (%)
        historical_baseline_velocity_ms:
          type: number
          nullable: true
          description: Load-matched historical mean-velocity baseline in m/s
        historical_comparison_data_points:
          type: integer
          description: Number of historical load-matched sessions used for the comparison
        diary_context_available:
          type: boolean
          description: Whether an optional diary sentiment context was available for this set date
        diary_context:
          type: string
          nullable: true
          description: Non-clinical sentiment context only; does not contain the raw diary note
      required:
        - status
        - exercise_name
        - mean_velocity_ms
        - peak_velocity_ms
        - first_rep_peak_ms
        - estimated_1rm_pct
        - velocity_zone
        - velocity_loss_pct
        - fatigue_level
        - sample_count
        - duration_s
        - ai_feedback
        - sparkden_history_used
        - cns_readiness_score
        - motor_readiness_level
        - velocity_trend
        - readiness_data_points
        - baseline_velocity_ms
        - actual_reps
        - rep_peaks_ms
        - historical_comparison
        - historical_comparison_delta_pct
        - historical_baseline_velocity_ms
        - historical_comparison_data_points
        - diary_context_available
        - diary_context
    DemoHistoricalSet:
      type: object
      description: Immutable demo history row. Loads are always canonical kilograms.
      properties:
        id:
          type: string
        setNumber:
          type: integer
        exercise:
          type: string
        daysBeforeDemo:
          type: integer
        loadKg:
          type: number
          description: Canonical load in kilograms
        targetReps:
          type: integer
        actualReps:
          type: integer
        meanRepTimeSec:
          type: number
        velocityLossPct:
          type: number
        displacementM:
          type: number
          nullable: true
        source:
          type: string
        provenance:
          type: string
      required:
        - id
        - setNumber
        - exercise
        - daysBeforeDemo
        - loadKg
        - targetReps
        - actualReps
        - meanRepTimeSec
        - velocityLossPct
        - displacementM
        - source
        - provenance