# lib/api-zod/src/generated/types/diaryEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiaryEntrySentiment } from './diaryEntrySentiment';
import type { DiaryEntrySentimentStatus } from './diaryEntrySentimentStatus';

export interface DiaryEntry {
  id: string;
  entry_date: Date;
  note: string;
  sentiment: DiaryEntrySentiment;
  sentiment_confidence: number | null;
  /** Short non-clinical summary; never a diagnosis */
  sentiment_summary: string | null;
  sentiment_status: DiaryEntrySentimentStatus;
  updated_at: Date;
}