# lib/api-zod/src/generated/types/diaryEntryUpsert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiaryEntryUpsert {
  /**
     * Open-text training, mood, recovery, or life note for the selected day
     * @minLength 1
     * @maxLength 2000
     */
  note: string;
}