# lib/api-zod/src/generated/types/diaryTrend.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiaryTrendPoint } from './diaryTrendPoint';

export interface DiaryTrend {
  days: number;
  analyzed_entries: number;
  performance_data_points: number;
  correlation_summary: string;
  points: DiaryTrendPoint[];
}