# lib/api-zod/src/generated/types/diaryTrendPoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiaryTrendPointSentiment } from './diaryTrendPointSentiment';

export interface DiaryTrendPoint {
  entry_date: Date;
  sentiment: DiaryTrendPointSentiment;
  mean_velocity_ms: number | null;
}