# lib/api-zod/src/generated/types/setAnalysisRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccelerationSample } from './accelerationSample';
import type { SetAnalysisRequestDisplayUnit } from './setAnalysisRequestDisplayUnit';
import type { SetAnalysisRequestPhonePlacement } from './setAnalysisRequestPhonePlacement';

/**
 * Exercise setup metadata and the raw sensor batch captured during a set
 */
export interface SetAnalysisRequest {
  /**
     * Name of the exercise (e.g. "Bench Press")
     * @minLength 1
     */
  exercise_name: string;
  /** Load on the bar in kilograms */
  weight_kg: number;
  /** User-facing mass unit used for load labels and coaching copy */
  display_unit?: SetAnalysisRequestDisplayUnit;
  /** User-entered load in display_unit; weight_kg remains canonical */
  display_load?: number;
  /** Planned rep count for the set */
  target_reps: number;
  /** Total sets planned in the workout */
  total_sets: number;
  /** Ordered raw acceleration samples captured during the set */
  samples: AccelerationSample[];
  /** How the phone is mounted during the set. Affects which axis is used for velocity integration and how the AI coaching interprets the data. */
  phone_placement?: SetAnalysisRequestPhonePlacement;
}